requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0

# MCP
mcp>=1.0.0

# Memory
chromadb>=0.5.0

# Search
tavily-python>=0.3.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Frontend
gradio>=5.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
Pillow>=10.0.0
aiofiles>=23.0.0
